add auth

+++ api/helper.py
@@ -37,13 +37,31 @@ def get_specific(self, request, pk):
 
 def put_specific(self, request, pk):    
     data = self.get_queryset(pk)
-    serializer = self.serializer_class(data, data=request.data)
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    if(request.user == data.creator): # If creator is who makes request
+        serializer = self.serializer_class(data, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    else:
+        content = {
+            'status': 'UNAUTHORIZED'
+        }
+        return Response(content, status=status.HTTP_401_UNAUTHORIZED)    
+        
 
 def delete_specific(self, request, pk, format=None):
     data = self.get_queryset(pk)
-    data.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT) 
+
+    if(request.user == data.creator): # If creator is who makes request
+        data.delete()
+        content = {
+            'status': 'NO CONTENT'
+        }
+        return Response(content, status=status.HTTP_204_NO_CONTENT)
+    else:
+        content = {
+            'status': 'UNAUTHORIZED'
+        }
+        return Response(content, status=status.HTTP_401_UNAUTHORIZED)
